Move e2e codemode builder to test helper
